+++ src/utils/nodesToNode.js
@@ -1,10 +1,13 @@
-export const nodesToNode = (nodes, document) => {
-  nodes = nodes.map(node => {
+export const nodesToNodes = (nodes, document) =>
+  nodes.map(node => {
     if (typeof node === 'string') {
       return document.createTextNode(node)
     }
     return node
   })
+
+export const nodesToNode = (nodes, document) => {
+  nodes = nodesToNodes(nodes, document)
   if (nodes.length === 1) {
     return nodes[0]
   }